Move test classes for navigation to success and failure packages
