+++ src/test/java/uk/gov/companieshouse/web/accounts/util/navigator/NavigatorTests.java
@@ -11,6 +11,15 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.web.servlet.view.UrlBasedViewResolver;
 import uk.gov.companieshouse.web.accounts.exception.MissingAnnotationException;
 import uk.gov.companieshouse.web.accounts.util.Navigator;
+import uk.gov.companieshouse.web.accounts.util.navigator.failure.MockControllerFive;
+import uk.gov.companieshouse.web.accounts.util.navigator.failure.MockControllerFour;
+import uk.gov.companieshouse.web.accounts.util.navigator.failure.MockControllerOne;
+import uk.gov.companieshouse.web.accounts.util.navigator.failure.MockControllerSeven;
+import uk.gov.companieshouse.web.accounts.util.navigator.failure.MockControllerThree;
+import uk.gov.companieshouse.web.accounts.util.navigator.failure.MockControllerTwo;
+import uk.gov.companieshouse.web.accounts.util.navigator.success.MockSuccessJourneyControllerOne;
+import uk.gov.companieshouse.web.accounts.util.navigator.success.MockSuccessJourneyControllerThree;
+import uk.gov.companieshouse.web.accounts.util.navigator.success.MockSuccessJourneyControllerTwo;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -36,7 +45,7 @@ public class NavigatorTests {
         Throwable exception = assertThrows(MissingAnnotationException.class, () ->
                 navigator.getNextControllerRedirect(MockControllerThree.class, COMPANY_NUMBER, TRANSACTION_ID, COMPANY_ACCOUNTS_ID));
 
-        assertEquals("Missing @NextController annotation on class uk.gov.companieshouse.web.accounts.util.navigator.MockControllerThree", exception.getMessage());
+        assertEquals("Missing @NextController annotation on class uk.gov.companieshouse.web.accounts.util.navigator.failure.MockControllerThree", exception.getMessage());
     }
 
     @Test
@@ -44,7 +53,7 @@ public class NavigatorTests {
         Throwable exception = assertThrows(MissingAnnotationException.class, () ->
                 navigator.getPreviousControllerPath(MockControllerThree.class, COMPANY_NUMBER, TRANSACTION_ID, COMPANY_ACCOUNTS_ID));
 
-        assertEquals("Missing @PreviousController annotation on class uk.gov.companieshouse.web.accounts.util.navigator.MockControllerThree", exception.getMessage());
+        assertEquals("Missing @PreviousController annotation on class uk.gov.companieshouse.web.accounts.util.navigator.failure.MockControllerThree", exception.getMessage());
     }
 
     @Test
@@ -52,7 +61,7 @@ public class NavigatorTests {
         Throwable exception = assertThrows(MissingAnnotationException.class, () ->
                 navigator.getNextControllerRedirect(MockControllerOne.class, COMPANY_NUMBER, TRANSACTION_ID, COMPANY_ACCOUNTS_ID));
 
-        assertEquals("Missing @RequestMapping annotation on class uk.gov.companieshouse.web.accounts.util.navigator.MockControllerTwo", exception.getMessage());
+        assertEquals("Missing @RequestMapping annotation on class uk.gov.companieshouse.web.accounts.util.navigator.failure.MockControllerTwo", exception.getMessage());
     }
 
     @Test
@@ -60,7 +69,7 @@ public class NavigatorTests {
         Throwable exception = assertThrows(MissingAnnotationException.class, () ->
                 navigator.getPreviousControllerPath(MockControllerTwo.class, COMPANY_NUMBER, TRANSACTION_ID, COMPANY_ACCOUNTS_ID));
 
-        assertEquals("Missing @RequestMapping annotation on class uk.gov.companieshouse.web.accounts.util.navigator.MockControllerOne", exception.getMessage());
+        assertEquals("Missing @RequestMapping annotation on class uk.gov.companieshouse.web.accounts.util.navigator.failure.MockControllerOne", exception.getMessage());
     }
 
     @Test
@@ -68,7 +77,7 @@ public class NavigatorTests {
         Throwable exception = assertThrows(MissingAnnotationException.class, () ->
                 navigator.getNextControllerRedirect(MockControllerFive.class, COMPANY_NUMBER, TRANSACTION_ID, COMPANY_ACCOUNTS_ID));
 
-        assertEquals("Missing @RequestMapping value on class uk.gov.companieshouse.web.accounts.util.navigator.MockControllerSix", exception.getMessage());
+        assertEquals("Missing @RequestMapping value on class uk.gov.companieshouse.web.accounts.util.navigator.failure.MockControllerSix", exception.getMessage());
     }
 
     @Test
@@ -76,7 +85,7 @@ public class NavigatorTests {
         Throwable exception = assertThrows(MissingAnnotationException.class, () ->
                 navigator.getPreviousControllerPath(MockControllerSeven.class, COMPANY_NUMBER, TRANSACTION_ID, COMPANY_ACCOUNTS_ID));
 
-        assertEquals("Missing @RequestMapping value on class uk.gov.companieshouse.web.accounts.util.navigator.MockControllerSix", exception.getMessage());
+        assertEquals("Missing @RequestMapping value on class uk.gov.companieshouse.web.accounts.util.navigator.failure.MockControllerSix", exception.getMessage());
     }
 
     @Test
