+++ src/test/java/uk/gov/companieshouse/web/accounts/util/navigator/failure/MockControllerFive.java
@@ -1,9 +1,10 @@
-package uk.gov.companieshouse.web.accounts.util.navigator;
+package uk.gov.companieshouse.web.accounts.util.navigator.failure;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import uk.gov.companieshouse.web.accounts.annotation.NextController;
 import uk.gov.companieshouse.web.accounts.annotation.PreviousController;
 import uk.gov.companieshouse.web.accounts.controller.BaseController;
+import uk.gov.companieshouse.web.accounts.util.navigator.NavigatorTests;
 
 /**
  * Mock controller class.
