+++ src/test/java/uk/gov/companieshouse/web/accounts/util/navigator/failure/MockControllerFour.java
@@ -1,18 +1,20 @@
-package uk.gov.companieshouse.web.accounts.util.navigator;
+package uk.gov.companieshouse.web.accounts.util.navigator.failure;
 
-import org.springframework.web.bind.annotation.RequestMapping;
 import uk.gov.companieshouse.web.accounts.annotation.NextController;
 import uk.gov.companieshouse.web.accounts.annotation.PreviousController;
 import uk.gov.companieshouse.web.accounts.controller.BaseController;
 import uk.gov.companieshouse.web.accounts.controller.ConditionalController;
+import uk.gov.companieshouse.web.accounts.util.navigator.NavigatorTests;
 
 /**
- * Mock controller class for success scenario testing of navigation.
+ * Mock conditional controller class for testing missing expected number of
+ * path variables.
+ *
+ * @see NavigatorTests
  */
-@NextController(MockSuccessJourneyControllerThree.class)
-@PreviousController(MockSuccessJourneyControllerOne.class)
-@RequestMapping("/mock-success-journey-controller-two")
-public class MockSuccessJourneyControllerTwo extends BaseController implements ConditionalController {
+@NextController(MockControllerFive.class)
+@PreviousController(MockControllerThree.class)
+public class MockControllerFour extends BaseController implements ConditionalController {
 
     @Override
     protected String getTemplateName() {
