+++ src/test/java/uk/gov/companieshouse/web/accounts/util/navigator/failure/MockControllerOne.java
@@ -1,18 +1,17 @@
-package uk.gov.companieshouse.web.accounts.util.navigator;
+package uk.gov.companieshouse.web.accounts.util.navigator.failure;
 
 import uk.gov.companieshouse.web.accounts.annotation.NextController;
-import uk.gov.companieshouse.web.accounts.annotation.PreviousController;
 import uk.gov.companieshouse.web.accounts.controller.BaseController;
+import uk.gov.companieshouse.web.accounts.util.navigator.NavigatorTests;
 
 /**
  * Mock controller class for testing missing navigation annotation {@code RequestMapping}
- * when attempting to obtain the next controller in the journey.
+ * when attempting to obtain the previous controller in the journey.
  *
  * @see NavigatorTests
  */
-@NextController(MockControllerThree.class)
-@PreviousController(MockControllerOne.class)
-public class MockControllerTwo extends BaseController {
+@NextController(MockControllerTwo.class)
+public class MockControllerOne extends BaseController {
 
     @Override
     protected String getTemplateName() {
