+++ src/test/java/uk/gov/companieshouse/web/accounts/util/navigator/failure/MockControllerSeven.java
@@ -1,7 +1,8 @@
-package uk.gov.companieshouse.web.accounts.util.navigator;
+package uk.gov.companieshouse.web.accounts.util.navigator.failure;
 
 import uk.gov.companieshouse.web.accounts.annotation.PreviousController;
 import uk.gov.companieshouse.web.accounts.controller.BaseController;
+import uk.gov.companieshouse.web.accounts.util.navigator.NavigatorTests;
 
 /**
  * Mock controller class for testing missing {@code RequestMapping} value
