+++ src/test/java/uk/gov/companieshouse/web/accounts/util/navigator/failure/MockControllerSix.java
@@ -1,8 +1,9 @@
-package uk.gov.companieshouse.web.accounts.util.navigator;
+package uk.gov.companieshouse.web.accounts.util.navigator.failure;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import uk.gov.companieshouse.web.accounts.annotation.PreviousController;
 import uk.gov.companieshouse.web.accounts.controller.BaseController;
+import uk.gov.companieshouse.web.accounts.util.navigator.NavigatorTests;
 
 /**
  * Mock controller class for testing missing {@code RequestMapping} value
